agents_backend/requirements.txt
beautifulsoup4==4.12.3
db-dtypes==1.3.1
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-aiplatform==1.81.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.13.0
google-cloud-storage==2.18.2
httpx==0.28.1
langchain==0.3.19
langchain-core==0.3.40
langchain-google-vertexai==2.0.14
langchain-text-splitters==0.3.6
langgraph==0.2.76
langgraph-checkpoint==2.0.16
langgraph-sdk==0.1.53
langsmith==0.3.11
nano-vectordb==0.0.4.3
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
pandas==2.2.3
passlib==1.7.4
plotly==5.24.1
PyJWT==2.8.0
SQLAlchemy==2.0.36
fastapi
fastapi-cli
